ui: node options: make text for `wakeonlan` option a bit more descriptive
If one is not familiar with the underlying configuration option,
one might think that one can enable Wake-On-Lan for a node with this
option. It was not really clear (at least without RTFM ;) ) that a
MAC address is supposed to be entered here. The added text should
hopefully make this a bit more clear.

Also increased the label width a bit, so that a full MAC
address can be properly displayed.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>